auto-implement refs #2617: Refactor Nyx relay notification delivery to channel sender plugins
